+++ ios/RNVIConstants.h
@@ -35,6 +35,15 @@ FOUNDATION_EXPORT NSString *const kEventEndpointRemoved;
 FOUNDATION_EXPORT NSString *const kEventEndpointVoiceActivityStarted;
 FOUNDATION_EXPORT NSString *const kEventEndpointVoiceActivityStopped;
 
+FOUNDATION_EXPORT NSString *const kEventQualityIssuePacketLoss;
+FOUNDATION_EXPORT NSString *const kEventQualityIssueCodecMismatch;
+FOUNDATION_EXPORT NSString *const kEventQualityIssueLocalVideoDegradation;
+FOUNDATION_EXPORT NSString *const kEventQualityIssueIceDisconnected;
+FOUNDATION_EXPORT NSString *const kEventQualityIssueHighMediaLatency;
+FOUNDATION_EXPORT NSString *const kEventQualityIssueNoAudioSignal;
+FOUNDATION_EXPORT NSString *const kEventQualityIssueNoAudioReceive;
+FOUNDATION_EXPORT NSString *const kEventQualityIssueNoVideoReceive;
+
 FOUNDATION_EXPORT NSString *const kEventAudioDeviceChanged;
 FOUNDATION_EXPORT NSString *const kEventAudioDeviceListChanged;
 
@@ -71,6 +80,15 @@ FOUNDATION_EXPORT NSString *const kEventNameEndpointRemoved;
 FOUNDATION_EXPORT NSString *const kEventNameVoiceActivityStarted;
 FOUNDATION_EXPORT NSString *const kEventNameVoiceActivityStopped;
 
+FOUNDATION_EXPORT NSString *const kEventNameQualityIssuePacketLoss;
+FOUNDATION_EXPORT NSString *const kEventNameQualityIssueCodecMismatch;
+FOUNDATION_EXPORT NSString *const kEventNameQualityIssueLocalVideoDegradation;
+FOUNDATION_EXPORT NSString *const kEventNameQualityIssueIceDisconnected;
+FOUNDATION_EXPORT NSString *const kEventNameQualityIssueHighMediaLatency;
+FOUNDATION_EXPORT NSString *const kEventNameQualityIssueNoAudioSignal;
+FOUNDATION_EXPORT NSString *const kEventNameQualityIssueNoAudioReceive;
+FOUNDATION_EXPORT NSString *const kEventNameQualityIssueNoVideoReceive;
+
 FOUNDATION_EXPORT NSString *const kEventEndpointStopReceivingVideoStreamSuccess;
 FOUNDATION_EXPORT NSString *const kEventEndpointStopReceivingVideoStreamFailure;
 FOUNDATION_EXPORT NSString *const kEventEndpointStartReceivingVideoStreamSuccess;
@@ -97,6 +115,12 @@ FOUNDATION_EXPORT NSString *const kEventParamCode;
 FOUNDATION_EXPORT NSString *const kEventParamMessage;
 FOUNDATION_EXPORT NSString *const kEventParamIncomingVideo;
 FOUNDATION_EXPORT NSString *const kEventParamCallKitUUID;
+FOUNDATION_EXPORT NSString *const kEventParamIssueLevel;
+FOUNDATION_EXPORT NSString *const kEventParamLatency;
+FOUNDATION_EXPORT NSString *const kEventParamActualSize;
+FOUNDATION_EXPORT NSString *const kEventParamTargetSize;
+FOUNDATION_EXPORT NSString *const kEventParamPacketLoss;
+FOUNDATION_EXPORT NSString *const kEventParamCodec;
 
 FOUNDATION_EXPORT NSString *const kEventParamLogLevel;
 FOUNDATION_EXPORT NSString *const kEventParamLogMessage;
@@ -120,6 +144,7 @@ FOUNDATION_EXPORT NSString *const kEventParamVideoStreamId;
 FOUNDATION_EXPORT NSString *const kEventParamCurrentAudioDevice;
 FOUNDATION_EXPORT NSString *const kEventParamDeviceList;
 FOUNDATION_EXPORT NSString *const kEventParamVideoStreamType;
+FOUNDATION_EXPORT NSString *const kEventParamAudioStreamId;
 
 FOUNDATION_EXPORT NSString *const kEventParamAudioFileId;
 FOUNDATION_EXPORT NSString *const kEventParamError;