+++ ios/RNVIConstants.m
@@ -35,6 +35,15 @@ NSString *const kEventEndpointRemoved = @"VIEndpointRemoved";
 NSString *const kEventEndpointVoiceActivityStarted = @"VIVoiceActivityStarted";
 NSString *const kEventEndpointVoiceActivityStopped = @"VIVoiceActivityStopped";
 
+NSString *const kEventQualityIssuePacketLoss = @"VIQualityIssuePacketLoss";
+NSString *const kEventQualityIssueCodecMismatch = @"VIQualityIssueCodecMismatch";
+NSString *const kEventQualityIssueLocalVideoDegradation = @"VIQualityIssueLocalVideoDegradation";
+NSString *const kEventQualityIssueIceDisconnected = @"VIQualityIssueIceDisconnected";
+NSString *const kEventQualityIssueHighMediaLatency = @"VIQualityIssueHighMediaLatency";
+NSString *const kEventQualityIssueNoAudioSignal = @"VIQualityIssueNoAudioSignal";
+NSString *const kEventQualityIssueNoAudioReceive = @"VIQualityIssueNoAudioReceive";
+NSString *const kEventQualityIssueNoVideoReceive = @"VIQualityIssueNoVideoReceive";
+
 NSString *const kEventEndpointStopReceivingVideoStreamSuccess = @"VIStopReceivingVideoStreamSuccess";
 NSString *const kEventEndpointStopReceivingVideoStreamFailure = @"VIStopReceivingVideoStreamFailure";
 NSString *const kEventEndpointStartReceivingVideoStreamSuccess = @"VIStartReceivingVideoStreamSuccess";
@@ -80,6 +89,15 @@ NSString *const kEventNameEndpointRemoved = @"Removed";
 NSString *const kEventNameVoiceActivityStarted = @"VoiceActivityStarted";
 NSString *const kEventNameVoiceActivityStopped = @"VoiceActivityStopped";
 
+NSString *const kEventNameQualityIssuePacketLoss = @"PacketLoss";
+NSString *const kEventNameQualityIssueCodecMismatch = @"CodecMismatch";
+NSString *const kEventNameQualityIssueLocalVideoDegradation = @"LocalVideoDegradation";
+NSString *const kEventNameQualityIssueIceDisconnected = @"IceDisconnected";
+NSString *const kEventNameQualityIssueHighMediaLatency = @"HighMediaLatency";
+NSString *const kEventNameQualityIssueNoAudioSignal = @"NoAudioSignal";
+NSString *const kEventNameQualityIssueNoAudioReceive = @"NoAudioReceive";
+NSString *const kEventNameQualityIssueNoVideoReceive = @"NoVideoReceive";
+
 NSString *const kEventNameAudioDeviceChanged = @"DeviceChanged";
 NSString *const kEventNameAudioDeviceListChanged = @"DeviceListChanged";
 
@@ -99,6 +117,12 @@ NSString *const kEventParamCode = @"code";
 NSString *const kEventParamMessage = @"message";
 NSString *const kEventParamIncomingVideo = @"video";
 NSString *const kEventParamCallKitUUID = @"callKitUUID";
+NSString *const kEventParamIssueLevel = @"issueLevel";
+NSString *const kEventParamLatency = @"latency";
+NSString *const kEventParamActualSize = @"actualSize";
+NSString *const kEventParamTargetSize = @"targetSize";
+NSString *const kEventParamPacketLoss = @"packetLoss";
+NSString *const kEventParamCodec = @"codec";
 
 NSString *const kEventParamLogLevel = @"level";
 NSString *const kEventParamLogMessage = @"message";
@@ -122,6 +146,7 @@ NSString *const kEventParamVideoStreamId = @"videoStreamId";
 NSString *const kEventParamCurrentAudioDevice = @"currentDevice";
 NSString *const kEventParamDeviceList = @"newDeviceList";
 NSString *const kEventParamVideoStreamType = @"videoStreamType";
+NSString *const kEventParamAudioStreamId = @"audioStreamId";
 
 NSString *const kEventParamAudioFileId = @"fileId";
 NSString *const kEventParamError = @"error";