+++ src/Voximplant.js
@@ -17,6 +17,7 @@ import VideoView from './call/VideoView';
 import EndpointEvents from './call/EndpointEvents';
 import ClientEvents from './client/ClientEvents';
 import {ClientState, CameraType, RenderScaleType, LogLevel, CallError, VideoCodec, RequestAudioFocusMode, VideoStreamType} from "./Enums";
+import QualityEvents from './call/QualityEvents';
 import * as Hardware from './hardware';
 import * as Messaging from './messaging';
 
@@ -58,6 +59,7 @@ export {
     VideoCodec,
     RequestAudioFocusMode,
     VideoStreamType,
+    QualityEvents,
     Hardware,
     Messaging
 };